seed update

+++ frontend/components/comment/comment_index_item.jsx
@@ -3,14 +3,21 @@ import React from 'react';
 const CommentIndexItem = ({comment, deleteComment}) => {
   // console.log(deleteComment, 'this is delete comment');
   // console.log(fetchComments, 'this is fetch comments');
+  let ownerDelete;
+  if (comment.cOwnerOwns) {
+    ownerDelete = (
+      <button onClick={()=> deleteComment(comment.id)} className="delete-comment-btn">
+        <i className="fas fa-times"></i>
+      </button>
+    );
+  }
+
   return (
     <div className="comment-index-item">
       <div className="cmt-username-row">
         {comment.username}
         <span>Posted {comment.updated} ago</span>
-        <button onClick={()=> deleteComment(comment.id)} className="delete-comment-btn">
-          <i className="fas fa-times"></i>
-        </button>
+        {ownerDelete}
       </div>
 
       <div className="cmt-body">
